refactor: submitPopUp
